pyproject.toml
[tool.poetry]
name = "pet-service"
version = "0.1.0"
description = ""
authors = ["Daniil Serov <zendrokalash1313@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.8"
loguru = "^0.6.0"
pandas = "^1.4.2"
uuid = "^1.30"
torch = "^2.2.1"
torchvision = "^0.17.1"
Flask = "^3.0.2"
gdown = "^5.1.0"
telebot = "^0.0.5"
pillow = "^10.2.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
